feat: reformat and split modules
