+++ .gitignore
@@ -6,4 +6,4 @@
 
 local.toml
 
-state.db
+*.db
