+++ circuits/plonky2x/src/circuits/mod.rs
@@ -1,0 +1,5 @@
+pub mod sync;
+pub mod verify;
+
+pub use sync::SyncCircuit;
+pub use verify::VerifyCircuit;
