+++ crates/rpc/src/prelude.rs
@@ -1,13 +1,13 @@
-pub use anyhow::anyhow;
-pub use anyhow::Result;
-pub use futures::FutureExt;
-pub use futures::TryFutureExt;
+pub use anyhow::{anyhow, Result};
+pub use futures::{FutureExt, TryFutureExt};
 pub use itertools::Itertools;
 pub use log::{debug, error, info, trace, warn};
 pub use near_primitives::types::AccountId;
 use near_primitives::types::TransactionOrReceiptId;
-pub use near_primitives_core::borsh::{self, BorshDeserialize, BorshSerialize};
-pub use near_primitives_core::hash::CryptoHash;
+pub use near_primitives_core::{
+    borsh::{self, BorshDeserialize, BorshSerialize},
+    hash::CryptoHash,
+};
 pub use serde::{Deserialize, Serialize};
 
 pub type Header = near_primitives::views::LightClientBlockLiteView;
